modularized Atum: S3Location moved to model, test suite, too.
